snapcraft.yaml: vendor mtools
There is a regression in mtools 4.0.32 which is resolved in future
releases. This regression involves mcopy failing just in case target
directories don't exist when recursively copying files.

This fix is available in 23.10 but not 22.04, which means that when
this snap changes from the core22 base to the core24 base mtools can be
re-added as a stage-package and this mtools part can be removed.

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>