go: use a different snapd
When ubuntu-image creates an Ubuntu Core image, it executes some
commands from a snapd it fetches at build time (like PrepareImage).
This means that if you change something like managed bootloader assets
provided by and produced by snapd as part of this process, you will need
to use a different snapd to build ubuntu-image to take in those changes
and build the correct image.

This commit changes the upstream snapd to one based off a similar commit
as the currently targetted one, but instead targets riscv64 and adds
a couple of patches to add support for riscv64 to the grub bootloader
assets.

This commit should not be merged into ubuntu-image but rather, these
relevant changes should be:
1) upstreamed into snapd
2) backported to the relevant release of snapd used by ubuntu-image

Signed-off-by: Dilyn Corner <dilyn.corner@canonical.com>